backend/requirements.txt
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
numpy==2.4.4
opencv-python-headless==4.13.0.92
pillow==12.0.0
psycopg2-binary==2.9.11
pytesseract==0.3.13
python-dotenv==1.2.2
reportlab==4.4.4
requests==2.33.1
stripe==15.0.1
